Add a user CRD api
This adds a simple CRD (no update at the moment) API for creating users,
and stating which channels they are allowed to connect to.

When a user is created a corresponding UserServer will be started, which
can be queries at a later point.  This should client apps to control
whether or not they are allowed to join particular channels.