Remove sessions.
Sessions require us to configure a secret_key_base, which needs to be
populated for prod releases.  I don't really want to have to do that,
and have no need for sessions, so removing them for now.

This should fix building of releases on circle, which was failing due to
a missing prod.secret.exs config file.  This file is no longer required.